code cleanup for types
